pyproject.toml
[tool.poetry]
name = "decryptpdf"
version = "0.1.0"
description = ""
authors = ["TomohiroHiratsuka"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
